@@ -3,7 +3,7 @@ description: A new Flutter application.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -18,24 +18,23 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  velocity_x: ^0.3.4
-  google_fonts: ^1.1.0
+  velocity_x: ^2.3.1-nullsafety.0
+  google_fonts: ^2.0.0
   flutter_icons: ^1.1.0
-  url_launcher: ^5.4.7
-  responsive_builder: ^0.1.2
-  get_it: ^4.0.2
-  font_awesome_flutter: ^8.8.1
+  url_launcher: ^6.0.2
+  responsive_builder: ^0.4.0-nullsafety.1
+  get_it: ^6.0.0
+  font_awesome_flutter: ^9.0.0-nullsafety
   steps: ^1.0.1
-
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.2
 
 dev_dependencies:
   flutter_test:
@@ -46,7 +45,6 @@ dev_dependencies:
 
 # The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
