@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
 class TranslateOnHover extends StatefulWidget {
-  final Widget child;
+  final Widget? child;
   // You can also pass the translation in here if you want to
-  TranslateOnHover({Key key, this.child}) : super(key: key);
+  TranslateOnHover({Key? key, this.child}) : super(key: key);
 
   @override
   _TranslateOnHoverState createState() => _TranslateOnHoverState();
