@@ -7,7 +7,7 @@ import 'package:steps/steps.dart';
 import 'package:velocity_x/velocity_x.dart';
 
 class ExperienceViewDesktop extends StatelessWidget {
-  const ExperienceViewDesktop({Key key}) : super(key: key);
+  const ExperienceViewDesktop({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -63,7 +63,7 @@ class ExperienceViewDesktop extends StatelessWidget {
 
 class ExperiencePictureWidget extends StatelessWidget {
   const ExperiencePictureWidget({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
