@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:steps/steps.dart';
 
 class ExperiencesDesktop extends StatelessWidget {
-  const ExperiencesDesktop({Key key}) : super(key: key);
+  const ExperiencesDesktop({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
