@@ -5,7 +5,7 @@ import 'package:satyam_website/widgets/method/method.dart';
 import 'package:velocity_x/velocity_x.dart';
 
 class ProjectViewDesktop extends StatelessWidget {
-  const ProjectViewDesktop({Key key}) : super(key: key);
+  const ProjectViewDesktop({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
