@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:velocity_x/velocity_x.dart';
 
 class MyPicAboutMobile extends StatelessWidget {
-  const MyPicAboutMobile({Key key}) : super(key: key);
+  const MyPicAboutMobile({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -15,7 +15,7 @@ class MyPicAboutMobile extends StatelessWidget {
 
 class MyPicWidget extends StatelessWidget {
   const MyPicWidget({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
