@@ -24,7 +24,7 @@ class IntroductionAboutDesktop extends StatelessWidget {
 
 class Introduction extends StatelessWidget {
   const Introduction({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   Widget technology(BuildContext context, String text) {
