@@ -27,7 +27,7 @@ class IntroductionTabletDesktop extends StatelessWidget {
 
 class Introduction extends StatelessWidget {
   const Introduction({
-    Key key,
+    Key? key,
   }) : super(key: key);
 
   @override
