@@ -3,7 +3,7 @@ import 'package:satyam_website/views/home/name_widget/name_widget.dart';
 import 'package:satyam_website/views/home/Introduction/introduction.dart';
 
 class HomeContentMobile extends StatelessWidget {
-  const HomeContentMobile({Key key}) : super(key: key);
+  const HomeContentMobile({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
