@@ -5,7 +5,7 @@ import 'package:satyam_website/views/home/home_content_desktop.dart';
 
 
 class HomeView extends StatelessWidget {
-  const HomeView({Key key}) : super(key: key);
+  const HomeView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
