@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 class CenteredView extends StatelessWidget {
-  final Widget child;
-  const CenteredView({Key key, this.child}) : super(key: key);
+  final Widget? child;
+  const CenteredView({Key? key, this.child}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
