@@ -1,22 +1,22 @@
 import 'package:flutter/material.dart';
 
 class CustomText extends StatelessWidget {
-  final String text;
+  final String? text;
   final double textsize;
   final Color color;
-  final double letterSpacing;
-  final FontWeight fontWeight;
+  final double? letterSpacing;
+  final FontWeight? fontWeight;
   CustomText(
-      {@required this.text,
-      @required this.textsize,
-      @required this.color,
+      {required this.text,
+      required this.textsize,
+      required this.color,
       this.letterSpacing,
       this.fontWeight
     });
   @override
   Widget build(BuildContext context) {
     return Text(
-      text,
+      text!,
       style: TextStyle(
           color: color,
           fontSize: textsize,
