@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class DrawerLogo extends StatelessWidget {
-  const DrawerLogo({Key key}) : super(key: key);
+  const DrawerLogo({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
