@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 
 class NavBarLogo extends StatelessWidget {
-  const NavBarLogo({Key key}) : super(key: key);
+  const NavBarLogo({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
