@@ -5,18 +5,18 @@ import 'package:satyam_website/extensions/translate_on_hover.dart';
 extension HoverExtensions on Widget {
   // Get a regerence to the body of the view
   static final appContainer =
-  html.window.document.getElementById('app-container');
+      html.window.document.getElementById('app-container');
 
   Widget get showCursorOnHover {
     return MouseRegion(
       child: this,
       // When the mouse enters the widget set the cursor to pointer
       onHover: (event) {
-        appContainer.style.cursor = 'pointer';
+        appContainer!.style.cursor = 'pointer';
       },
       // When it exits set it back to default
       onExit: (event) {
-        appContainer.style.cursor = 'default';
+        appContainer!.style.cursor = 'default';
       },
     );
   }
